add snapshot testing (#3)
Added Jest snapshot testing. This help when refactoring the code but we don't expect changes in the infrastructure. If the snapshot matches the previous state, then we are certain that no changes will happen in the infrastructure.

Had to work-around some cdk/jest issues. Docker image assets are not reproducible, and had to omit them from the snapshot, so that the snapshot is the same when created on a user pc and on a Github CI.